[palo-alto-cortex-xdr] old code
